renamed go client dir to 'go'
