+++ cmd/aurum/commands.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/finitum/aurum/clients/aurum-go"
+	"github.com/finitum/aurum/clients/go"
 	"github.com/finitum/aurum/pkg/jwt"
 	"github.com/finitum/aurum/pkg/models"
 )
